updates to README for v2.0.0 (#183)
For some reason, I thought I'd done this with the first beta release. I
need to link to this (esp the upgrading section) in the beta release dev
docs.

There are a couple of options here. We could either:
- Merge with a disclaimer at the top that this is the beta README and to
go to `<insert link>` for version one.
- Keep this on a branch. Looks like I can still [deep
link](https://github.com/fullstorydev/fullstory-browser-sdk/blob/scottnorvell/v2-readme-updates/README.md#migrating-to-version-200)
to the README from the branch and it is available to the public. (at
least an incognito window)

Thoughts on the above?

---------

Co-authored-by: Patrick@FullStory <45576380+patrick-fs@users.noreply.github.com>